requirements.txt
pandas
psycopg2-binary
pyyaml
python-dotenv